// src/app/interfaces/program-review.ts
export interface ProgramReview {
    reviewerUid: string;
    reviewerFirstName: string;
    reviewerLastName: string;
    sellerUid: string;
    programId: string;
    starValue: number;
    lastUpdated: number; // unix timestamp
    summary?: string;
    summaryExists?: boolean; // set true if review includes summary text
    reviewerPhoto?: string;
}